Added -e/--manual-error-cuttoffs option
User is now able to specify a list of error curve cutoffs, one for each
peak in a simiar manner to the repeat k-mer cutoffs. The k-mers included
in this error curve will not be counted when we are counting the total
number of k-mers present. We will see if this affects the GRI in any
meaningful way.