Corrected and renamed find_main_peak()
I noticted that I had an error in my implementation of how to find the x
value of the first peak. Basically, it used to be the case that I could
think that other peaks occurring before the main peak (the heterozygous
peak, for instance) were in fact the main peak. I have therefore
rewritten my find_first_peak_max function in order to hopefully fix
this, and have also renamed it to find_main_peak because I think that
this name is more descriptive.