Changed repeat k-mer cutoff prediction function
I have found that multiplying the x value for the main peak by 1.65 and
adding 1 seems to give a good prediction of where the repetitive k-mers
start to occur (judged on spectra with nice peaks). I have therefore
implemented this new method, but left the old one commented out as a
reference.