Added -u/--upper-bound option
This option provides an upper limit on the number of occurrences a k-mer
can have before it no longer counts towards the total number of k-mers or
the number of repeat k-mers. This is in an effort to see whether we can
reduce the effect of PCR duplication, and to see whether this makes our
calculated GRIs more consistnent within the same genome, as we should be
able to expect.