Chnaged how I find start of first peak
I now return the smallest minimum, instead of the first minimum
preceding the peak. This change was necessary because I was often
returning the minimum between the homozygous and the heterozygous peak
as the start of the main peak, which messes up my method for finding the
start of the repetitive k-mers.