Use 1 instead of 0 as min cutoff if unspecified
Previously I was using 0 to as the minimum occurrence cutoff if it had
been unspecified by the user. However, this was below the minumum value
of 1 required by the parser for this value, so it seemed silly to have
it used here. Therefore, I've changed it to 1, which has the same
effect.