Return actual x val not its index in extrema lists
Previously, we were returning the index of the extremum in the extrema
list functions, instead of, as would have been correct, returning the
occurrence with this index. This, generally, would only lead to an
off-by-one error (hist dicts normally contain keys sequentially
increasing from one, and therefore a key with index n would generally
correspond to an occurrence number of (n+1)). However: (1) This cannot
be guaranteed; and (2) This was incorrect in the first place, and not
the intended behaviour. I feel that I fixed this a little while ago, but
can't find any record of that fix now... Anyway, these two functions now
return the actual occurrence values at the extrema, rather than the
indices of these values.