Use current error cutoff in repeat cutoff calc
Before, I was predicting a new error cutoff in my repeat cutoff
estimation function, using the same method as I do in my error cutoff
estimation function. However, this is not the best way to do it, as I
may choose how I estimate the error cutoff. Also, it does not use the
error cutoff if specfied by the user. Now, the function just takes the
error cutoff as an argument, thus solving both of these problems.