Changed import mechanism for optional modules
Numpy, Scipy, and Matplotlib are not required for all functions of GRIn.
Therefore they should not cause it to crash should they not be required.
Now, these three modules are imported globally, but will not cause a
crash if they are cannot be imported if they are not required. Before, I
was doing this by importing them several times in the functions where
they are needed.