Changed --verbose flag to --verbosity
This new flag counts the number of times it was used in the command line
arguments. A value of 1 will trigger more textual output; a value of 2
will trigger more textual output and the k-mer spectra to be displayed
with the cutoffs, meaning that the user can see where the cutoffs are
set.