Apply default verbosity of 0 to args.verbosity
If the verbosity flag is not set by the user, argparse sets it to None.
This leads to problems when comparing it to 0 in Python 3. Before, I was
only actually changing my renamed verbosity setting (i.e. the variable
set by 'verbosity = args.verbosity'). With this change, I am now
applying this default setting of 0 to args.verbosity itself. This allows
args to be passed around without creating problems.