Increase difference_cutoff estimation failed
Previously, I was using just one value of difference_cutoff (the cutoff
at which the mean difference between neighbouring points in a window was
deemed small enough to return the midpoint of that window as the Upper
Cutoff. Now, I increase differene_cutoff from 0 up to a maximum of 5
until I find an Upper Cutoff. If this still fails, I return 20 * the
k-mer depth as the Upper Cutoff.